move inline test into test/
