+++ example/convex/_generated/api.d.ts
@@ -12,11 +12,11 @@ import type * as admin from "../admin.js";
 import type * as catchError from "../catchError.js";
 import type * as e2e from "../e2e.js";
 import type * as example from "../example.js";
-import type * as inlineTest from "../inlineTest.js";
 import type * as nestedWorkflow from "../nestedWorkflow.js";
 import type * as oversized from "../oversized.js";
 import type * as passingSignals from "../passingSignals.js";
 import type * as test_contextRoundtrip from "../test/contextRoundtrip.js";
+import type * as test_inline from "../test/inline.js";
 import type * as test_oldSyntax from "../test/oldSyntax.js";
 import type * as transcription from "../transcription.js";
 import type * as userConfirmation from "../userConfirmation.js";
@@ -32,11 +32,11 @@ declare const fullApi: ApiFromModules<{
   catchError: typeof catchError;
   e2e: typeof e2e;
   example: typeof example;
-  inlineTest: typeof inlineTest;
   nestedWorkflow: typeof nestedWorkflow;
   oversized: typeof oversized;
   passingSignals: typeof passingSignals;
   "test/contextRoundtrip": typeof test_contextRoundtrip;
+  "test/inline": typeof test_inline;
   "test/oldSyntax": typeof test_oldSyntax;
   transcription: typeof transcription;
   userConfirmation: typeof userConfirmation;
